resume_builder/requirements.txt
cohere
cohere
gradio
jinja2
weasyprint
python-dotenv
